Cleanup some unnecessary code/fallbacks for dayamlchecker, codemirror
