+++ docassemble/ALWeaver/data/templates/editor.html
@@ -324,11 +324,8 @@
   window.__EDITOR_BOOTSTRAP__ = __EDITOR_BOOTSTRAP_JSON__;
 </script>
 <script src="/static/app/cm6.min.js"></script>
-<script src="/al/editor/static/editor_source_adapter.js"></script>
 <script src="/al/editor/static/editor_serializers.js"></script>
 <script src="/al/editor/static/editor_api_client.js"></script>
-<script src="/al/editor/static/editor_state_store.js"></script>
-<script src="/al/editor/static/editor_command_manager.js"></script>
 <script src="/al/editor/static/editor_dirty_state.js"></script>
 <script src="/al/editor/static/editor_validation_source.js"></script>
 <script src="/al/editor/static/editor_runtime_inspector.js"></script>
